+++ api/src/database/run-ast/lib/apply-query/filter.ts
@@ -1,0 +1,497 @@
+import { InvalidQueryError } from '@directus/errors';
+import type {
+	ClientFilterOperator,
+	FieldFunction,
+	FieldOverview,
+	Filter,
+	Permission,
+	Relation,
+	SchemaOverview,
+	Type,
+} from '@directus/types';
+import { getFilterOperatorsForType, getFunctionsForType, getOutputTypeForFunction } from '@directus/utils';
+import type { Knex } from 'knex';
+import { getCases } from '../../../../permissions/modules/process-ast/lib/get-cases.js';
+import type { AliasMap } from '../../../../utils/get-column-path.js';
+import { getColumnPath } from '../../../../utils/get-column-path.js';
+import { getColumn } from '../../../../utils/get-column.js';
+import { getRelationInfo } from '../../../../utils/get-relation-info.js';
+import { parseFilterKey } from '../../../../utils/parse-filter-key.js';
+import { getHelpers } from '../../../helpers/index.js';
+import { getFilterPath } from './get-filter-path.js';
+import { addJoin } from './add-join.js';
+import { getOperation } from './get-operation.js';
+import applyQuery from './index.js';
+
+export function applyFilter(
+	knex: Knex,
+	schema: SchemaOverview,
+	rootQuery: Knex.QueryBuilder,
+	rootFilter: Filter,
+	collection: string,
+	aliasMap: AliasMap,
+	cases: Filter[],
+	permissions: Permission[],
+) {
+	const helpers = getHelpers(knex);
+	const relations: Relation[] = schema.relations;
+	let hasJoins = false;
+	let hasMultiRelationalFilter = false;
+
+	addJoins(rootQuery, rootFilter, collection);
+	addWhereClauses(knex, rootQuery, rootFilter, collection);
+
+	return { query: rootQuery, hasJoins, hasMultiRelationalFilter };
+
+	function addJoins(dbQuery: Knex.QueryBuilder, filter: Filter, collection: string) {
+		// eslint-disable-next-line prefer-const
+		for (let [key, value] of Object.entries(filter)) {
+			if (key === '_or' || key === '_and') {
+				// If the _or array contains an empty object (full permissions), we should short-circuit and ignore all other
+				// permission checks, as {} already matches full permissions.
+				if (key === '_or' && value.some((subFilter: Record<string, any>) => Object.keys(subFilter).length === 0)) {
+					// But only do so, if the value is not equal to `cases` (since then this is not permission related at all)
+					// or the length of value is 1, ie. only the empty filter.
+					// If the length is more than one it means that some items (and fields) might now be available, so
+					// the joins are required for the case/when construction.
+					if (value !== cases || value.length === 1) {
+						continue;
+					} else {
+						// Otherwise we can at least filter out all empty filters that would not add joins anyway
+						value = value.filter((subFilter: Record<string, any>) => Object.keys(subFilter).length > 0);
+					}
+				}
+
+				value.forEach((subFilter: Record<string, any>) => {
+					addJoins(dbQuery, subFilter, collection);
+				});
+
+				continue;
+			}
+
+			const filterPath = getFilterPath(key, value);
+
+			if (
+				filterPath.length > 1 ||
+				(!(key.includes('(') && key.includes(')')) && schema.collections[collection]?.fields[key]?.type === 'alias')
+			) {
+				const { hasMultiRelational, isJoinAdded } = addJoin({
+					path: filterPath,
+					collection,
+					knex,
+					schema,
+					relations,
+					rootQuery,
+					aliasMap,
+				});
+
+				if (!hasJoins) {
+					hasJoins = isJoinAdded;
+				}
+
+				if (!hasMultiRelationalFilter) {
+					hasMultiRelationalFilter = hasMultiRelational;
+				}
+			}
+		}
+	}
+
+	function addWhereClauses(
+		knex: Knex,
+		dbQuery: Knex.QueryBuilder,
+		filter: Filter,
+		collection: string,
+		logical: 'and' | 'or' = 'and',
+	) {
+		for (const [key, value] of Object.entries(filter)) {
+			if (key === '_or' || key === '_and') {
+				// If the _or array contains an empty object (full permissions), we should short-circuit and ignore all other
+				// permission checks, as {} already matches full permissions.
+				if (key === '_or' && value.some((subFilter: Record<string, any>) => Object.keys(subFilter).length === 0)) {
+					continue;
+				}
+
+				/** @NOTE this callback function isn't called until Knex runs the query */
+				dbQuery[logical].where((subQuery) => {
+					value.forEach((subFilter: Record<string, any>) => {
+						addWhereClauses(knex, subQuery, subFilter, collection, key === '_and' ? 'and' : 'or');
+					});
+				});
+
+				continue;
+			}
+
+			const filterPath = getFilterPath(key, value);
+
+			/**
+			 * For A2M fields, the path can contain an optional collection scope <field>:<scope>
+			 */
+			const pathRoot = filterPath[0]!.split(':')[0]!;
+
+			const { relation, relationType } = getRelationInfo(relations, collection, pathRoot);
+
+			const operation = getOperation(key, value);
+
+			if (!operation) continue;
+
+			const { operator: filterOperator, value: filterValue } = operation;
+
+			if (
+				filterPath.length > 1 ||
+				(!(key.includes('(') && key.includes(')')) && schema.collections[collection]?.fields[key]?.type === 'alias')
+			) {
+				if (!relation) continue;
+
+				if (relationType === 'o2m' || relationType === 'o2a') {
+					let pkField: Knex.Raw<any> | string = `${collection}.${
+						schema.collections[relation!.related_collection!]!.primary
+					}`;
+
+					if (relationType === 'o2a') {
+						pkField = knex.raw(getHelpers(knex).schema.castA2oPrimaryKey(), [pkField]);
+					}
+
+					const childKey = Object.keys(value)?.[0];
+
+					if (childKey === '_none' || childKey === '_some') {
+						const subQueryBuilder =
+							(filter: Filter, cases: Filter[]) => (subQueryKnex: Knex.QueryBuilder<any, unknown[]>) => {
+								const field = relation!.field;
+								const collection = relation!.collection;
+								const column = `${collection}.${field}`;
+
+								subQueryKnex
+									.select({ [field]: column })
+									.from(collection)
+									.whereNotNull(column);
+
+								applyQuery(knex, relation!.collection, subQueryKnex, { filter }, schema, cases, permissions);
+							};
+
+						const { cases: subCases } = getCases(relation!.collection, permissions, []);
+
+						if (childKey === '_none') {
+							dbQuery[logical].whereNotIn(
+								pkField as string,
+								subQueryBuilder(Object.values(value)[0] as Filter, subCases),
+							);
+
+							continue;
+						} else if (childKey === '_some') {
+							dbQuery[logical].whereIn(pkField as string, subQueryBuilder(Object.values(value)[0] as Filter, subCases));
+							continue;
+						}
+					}
+				}
+
+				if (filterPath.includes('_none') || filterPath.includes('_some')) {
+					throw new InvalidQueryError({
+						reason: `"${
+							filterPath.includes('_none') ? '_none' : '_some'
+						}" can only be used with top level relational alias field`,
+					});
+				}
+
+				const { columnPath, targetCollection, addNestedPkField } = getColumnPath({
+					path: filterPath,
+					collection,
+					relations,
+					aliasMap,
+					schema,
+				});
+
+				if (addNestedPkField) {
+					filterPath.push(addNestedPkField);
+				}
+
+				if (!columnPath) continue;
+
+				const { type, special } = getFilterType(
+					schema.collections[targetCollection]!.fields,
+					filterPath.at(-1)!,
+					targetCollection,
+				)!;
+
+				validateFilterOperator(type, filterOperator, special);
+
+				applyFilterToQuery(columnPath, filterOperator, filterValue, logical, targetCollection);
+			} else {
+				const { type, special } = getFilterType(schema.collections[collection]!.fields, filterPath[0]!, collection)!;
+
+				validateFilterOperator(type, filterOperator, special);
+
+				const aliasedCollection = aliasMap['']?.alias || collection;
+
+				applyFilterToQuery(`${aliasedCollection}.${filterPath[0]}`, filterOperator, filterValue, logical, collection);
+			}
+		}
+
+		function getFilterType(fields: Record<string, FieldOverview>, key: string, collection = 'unknown') {
+			const { fieldName, functionName } = parseFilterKey(key);
+
+			const field = fields[fieldName];
+
+			if (!field) {
+				throw new InvalidQueryError({ reason: `Invalid filter key "${key}" on "${collection}"` });
+			}
+
+			const { type } = field;
+
+			if (functionName) {
+				const availableFunctions: string[] = getFunctionsForType(type);
+
+				if (!availableFunctions.includes(functionName)) {
+					throw new InvalidQueryError({ reason: `Invalid filter key "${key}" on "${collection}"` });
+				}
+
+				const functionType = getOutputTypeForFunction(functionName as FieldFunction);
+
+				return { type: functionType };
+			}
+
+			return { type, special: field.special };
+		}
+
+		function validateFilterOperator(type: Type, filterOperator: string, special?: string[]) {
+			if (filterOperator.startsWith('_')) {
+				filterOperator = filterOperator.slice(1);
+			}
+
+			if (!getFilterOperatorsForType(type).includes(filterOperator as ClientFilterOperator)) {
+				throw new InvalidQueryError({
+					reason: `"${type}" field type does not contain the "_${filterOperator}" filter operator`,
+				});
+			}
+
+			if (
+				special?.includes('conceal') &&
+				!getFilterOperatorsForType('hash').includes(filterOperator as ClientFilterOperator)
+			) {
+				throw new InvalidQueryError({
+					reason: `Field with "conceal" special does not allow the "_${filterOperator}" filter operator`,
+				});
+			}
+		}
+
+		function applyFilterToQuery(
+			key: string,
+			operator: string,
+			compareValue: any,
+			logical: 'and' | 'or' = 'and',
+			originalCollectionName?: string,
+		) {
+			const [table, column] = key.split('.');
+
+			// Is processed through Knex.Raw, so should be safe to string-inject into these where queries
+			const selectionRaw = getColumn(knex, table!, column!, false, schema, { originalCollectionName }) as any;
+
+			// Knex supports "raw" in the columnName parameter, but isn't typed as such. Too bad..
+			// See https://github.com/knex/knex/issues/4518 @TODO remove as any once knex is updated
+
+			// These operators don't rely on a value, and can thus be used without one (eg `?filter[field][_null]`)
+			if (
+				(operator === '_null' && compareValue !== false) ||
+				(operator === '_nnull' && compareValue === false) ||
+				(operator === '_eq' && compareValue === null)
+			) {
+				dbQuery[logical].whereNull(selectionRaw);
+				return;
+			}
+
+			if (
+				(operator === '_nnull' && compareValue !== false) ||
+				(operator === '_null' && compareValue === false) ||
+				(operator === '_neq' && compareValue === null)
+			) {
+				dbQuery[logical].whereNotNull(selectionRaw);
+				return;
+			}
+
+			if ((operator === '_empty' && compareValue !== false) || (operator === '_nempty' && compareValue === false)) {
+				dbQuery[logical].andWhere((query) => {
+					query.whereNull(key).orWhere(key, '=', '');
+				});
+			}
+
+			if ((operator === '_nempty' && compareValue !== false) || (operator === '_empty' && compareValue === false)) {
+				dbQuery[logical].andWhere((query) => {
+					query.whereNotNull(key).andWhere(key, '!=', '');
+				});
+			}
+
+			// The following fields however, require a value to be run. If no value is passed, we
+			// ignore them. This allows easier use in GraphQL, where you wouldn't be able to
+			// conditionally build out your filter structure (#4471)
+			if (compareValue === undefined) return;
+
+			if (Array.isArray(compareValue)) {
+				// Tip: when using a `[Type]` type in GraphQL, but don't provide the variable, it'll be
+				// reported as [undefined].
+				// We need to remove any undefined values, as they are useless
+				compareValue = compareValue.filter((val) => val !== undefined);
+			}
+
+			// Cast filter value (compareValue) based on function used
+			if (column!.includes('(') && column!.includes(')')) {
+				const functionName = column!.split('(')[0] as FieldFunction;
+				const type = getOutputTypeForFunction(functionName);
+
+				if (['integer', 'float', 'decimal'].includes(type)) {
+					compareValue = Array.isArray(compareValue) ? compareValue.map(Number) : Number(compareValue);
+				}
+			}
+
+			// Cast filter value (compareValue) based on type of field being filtered against
+			const [collection, field] = key.split('.');
+			const mappedCollection = (originalCollectionName || collection)!;
+
+			if (mappedCollection! in schema.collections && field! in schema.collections[mappedCollection]!.fields) {
+				const type = schema.collections[mappedCollection]!.fields[field!]!.type;
+
+				if (['date', 'dateTime', 'time', 'timestamp'].includes(type)) {
+					if (Array.isArray(compareValue)) {
+						compareValue = compareValue.map((val) => helpers.date.parse(val));
+					} else {
+						compareValue = helpers.date.parse(compareValue);
+					}
+				}
+
+				if (['integer', 'float', 'decimal'].includes(type)) {
+					if (Array.isArray(compareValue)) {
+						compareValue = compareValue.map((val) => Number(val));
+					} else {
+						compareValue = Number(compareValue);
+					}
+				}
+			}
+
+			if (operator === '_eq') {
+				dbQuery[logical].where(selectionRaw, '=', compareValue);
+			}
+
+			if (operator === '_neq') {
+				dbQuery[logical].whereNot(selectionRaw, compareValue);
+			}
+
+			if (operator === '_ieq') {
+				dbQuery[logical].whereRaw(`LOWER(??) = ?`, [selectionRaw, `${compareValue.toLowerCase()}`]);
+			}
+
+			if (operator === '_nieq') {
+				dbQuery[logical].whereRaw(`LOWER(??) <> ?`, [selectionRaw, `${compareValue.toLowerCase()}`]);
+			}
+
+			if (operator === '_contains') {
+				dbQuery[logical].where(selectionRaw, 'like', `%${compareValue}%`);
+			}
+
+			if (operator === '_ncontains') {
+				dbQuery[logical].whereNot(selectionRaw, 'like', `%${compareValue}%`);
+			}
+
+			if (operator === '_icontains') {
+				dbQuery[logical].whereRaw(`LOWER(??) LIKE ?`, [selectionRaw, `%${compareValue.toLowerCase()}%`]);
+			}
+
+			if (operator === '_nicontains') {
+				dbQuery[logical].whereRaw(`LOWER(??) NOT LIKE ?`, [selectionRaw, `%${compareValue.toLowerCase()}%`]);
+			}
+
+			if (operator === '_starts_with') {
+				dbQuery[logical].where(key, 'like', `${compareValue}%`);
+			}
+
+			if (operator === '_nstarts_with') {
+				dbQuery[logical].whereNot(key, 'like', `${compareValue}%`);
+			}
+
+			if (operator === '_istarts_with') {
+				dbQuery[logical].whereRaw(`LOWER(??) LIKE ?`, [selectionRaw, `${compareValue.toLowerCase()}%`]);
+			}
+
+			if (operator === '_nistarts_with') {
+				dbQuery[logical].whereRaw(`LOWER(??) NOT LIKE ?`, [selectionRaw, `${compareValue.toLowerCase()}%`]);
+			}
+
+			if (operator === '_ends_with') {
+				dbQuery[logical].where(key, 'like', `%${compareValue}`);
+			}
+
+			if (operator === '_nends_with') {
+				dbQuery[logical].whereNot(key, 'like', `%${compareValue}`);
+			}
+
+			if (operator === '_iends_with') {
+				dbQuery[logical].whereRaw(`LOWER(??) LIKE ?`, [selectionRaw, `%${compareValue.toLowerCase()}`]);
+			}
+
+			if (operator === '_niends_with') {
+				dbQuery[logical].whereRaw(`LOWER(??) NOT LIKE ?`, [selectionRaw, `%${compareValue.toLowerCase()}`]);
+			}
+
+			if (operator === '_gt') {
+				dbQuery[logical].where(selectionRaw, '>', compareValue);
+			}
+
+			if (operator === '_gte') {
+				dbQuery[logical].where(selectionRaw, '>=', compareValue);
+			}
+
+			if (operator === '_lt') {
+				dbQuery[logical].where(selectionRaw, '<', compareValue);
+			}
+
+			if (operator === '_lte') {
+				dbQuery[logical].where(selectionRaw, '<=', compareValue);
+			}
+
+			if (operator === '_in') {
+				let value = compareValue;
+				if (typeof value === 'string') value = value.split(',');
+
+				dbQuery[logical].whereIn(selectionRaw, value as string[]);
+			}
+
+			if (operator === '_nin') {
+				let value = compareValue;
+				if (typeof value === 'string') value = value.split(',');
+
+				dbQuery[logical].whereNotIn(selectionRaw, value as string[]);
+			}
+
+			if (operator === '_between') {
+				let value = compareValue;
+				if (typeof value === 'string') value = value.split(',');
+
+				if (value.length !== 2) return;
+
+				dbQuery[logical].whereBetween(selectionRaw, value);
+			}
+
+			if (operator === '_nbetween') {
+				let value = compareValue;
+				if (typeof value === 'string') value = value.split(',');
+
+				if (value.length !== 2) return;
+
+				dbQuery[logical].whereNotBetween(selectionRaw, value);
+			}
+
+			if (operator == '_intersects') {
+				dbQuery[logical].whereRaw(helpers.st.intersects(key, compareValue));
+			}
+
+			if (operator == '_nintersects') {
+				dbQuery[logical].whereRaw(helpers.st.nintersects(key, compareValue));
+			}
+
+			if (operator == '_intersects_bbox') {
+				dbQuery[logical].whereRaw(helpers.st.intersects_bbox(key, compareValue));
+			}
+
+			if (operator == '_nintersects_bbox') {
+				dbQuery[logical].whereRaw(helpers.st.nintersects_bbox(key, compareValue));
+			}
+		}
+	}
+}
